fix changes requested in PR review moved pantry models to pantry app
