+++ waitress/app/models.py
@@ -275,18 +275,3 @@ class MealSession(models.Model):
         """
         date_today = timezone.now().date()
         return cls.objects.filter(date=date_today, status=True)
-
-
-class Pantry(models.Model):
-    """
-    Model representing the pantry service
-    """
-
-    id = models.AutoField(unique=True, primary_key=True)
-    user = models.ForeignKey(SlackUser, on_delete=models.CASCADE)
-    date = models.DateField(auto_now_add=True)
-
-    @classmethod
-    def is_tapped(cls, user_id):
-        service = cls.objects.filter(user_id=user_id, date=date.today()).first()
-        return service
