+++ waitress/settings/base.py
@@ -33,6 +33,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "app",
+    "pantry",
     "corsheaders",
     "drf_yasg",
 ]
